Add files via upload
Measures each Zone's length and width dimensions (feet-inches) based on Bounding Box, and writes it to a custom property. We use a Zone Label to display these dimensions in plan. The script takes a custom property called "Zone Angle" (user input) in order to calculate the dimensions correctly for rotated zones. We did not find a way to pull the rotation angle automatically, so it defaults to 0 degrees and is filled in by the user if different. The math formula breaks at 45 degrees (a compromise, since to fix this would require another user input).